// src/types/poem.ts
export type PoemData = {
  id: number;
  // 标题
  title?: string;
  // 作者
  author?: string;
  // 段落
  para?: string;
  paragraphs?: string;
  // 标签
  tags?: string;
  // 注释
  notes?: string;
  // 分组
  groups?: string;
  // 封面
  cover?: string;
  // 韵律
  rhythmic?: string;
  // 章节
  chapter?: string;
  // 内容
  content?: string;
  // 注释
  comment?: string;
  // 卷
  volume?: string;
  // 传记
  biography?: string;
  // 部分
  section?: string;
  // 前言
  prologue?: string;
  // 原文
  origin?: string;
  // 来源
  source?: string;
  // 译文
  translation?: string;
  // 是否AI优化过
  ai_optimized?: number;
  // 原始段落（优化前）
  original_paragraphs?: string;
  // 拼音
  pinyin?: string;
};

export type Poem = {
  id: number;
  // 标题
  title?: string;
  // 作者
  author?: string;
  // 段落
  para?: string[];
  paragraphs?: string[];
  // 标签
  tags?: string[];
  // 注释
  notes?: string[];
  // 分组
  groups?: string[];
  // 封面
  cover?: string;
  // 韵律
  rhythmic?: string;
  // 章节
  chapter?: string;
  // 内容
  content?: string;
  // 注释
  comment?: string[];
  // 卷
  volume?: string;
  // 传记
  biography?: string;
  // 部分
  section?: string;
  // 前言
  prologue?: string;
  // 原文
  origin?: string;
  // 来源
  source?: string;
  // 译文
  translation?: string;
  // 是否AI优化过
  ai_optimized?: number;
  // 原始段落（优化前）
  original_paragraphs?: string[];
  // 拼音
  pinyin?: string[];
}